implement architecture changes without tests
