+++ brewblox_devcon_spark/api/system_api.py
@@ -8,8 +8,8 @@ from aiohttp import web
 from aiohttp_apispec import docs, response_schema
 from brewblox_service import brewblox_logger, mqtt, scheduler, strex
 
-from brewblox_devcon_spark import (commander, device, exceptions,
-                                   service_status, ymodem)
+from brewblox_devcon_spark import (commander, exceptions, service_status,
+                                   spark, ymodem)
 from brewblox_devcon_spark.api import schemas
 
 TRANSFER_TIMEOUT_S = 30
@@ -70,7 +70,7 @@ class FirmwareUpdater():
 
     async def flash(self) -> dict:  # pragma: no cover
         sender = ymodem.FileSender(self._notify)
-        cmder = commander.get_commander(self.app)
+        cmder = commander.fget_commander(self.app)
         address = service_status.desc(self.app).device_address
 
         self._notify(f'Started updating {self.name}@{address} to version {self.version} ({self.date})')
@@ -116,7 +116,7 @@ class SystemApi():
     async def reboot(self):
         async def wrapper():
             try:
-                await device.get_device(self.app).reboot()
+                await spark.fget(self.app).reboot()
             except exceptions.CommandTimeout:
                 pass
             except Exception as ex:  # pragma: no cover
@@ -127,7 +127,7 @@ class SystemApi():
     async def factory_reset(self):
         async def wrapper():
             try:
-                await device.get_device(self.app).factory_reset()
+                await spark.fget(self.app).factory_reset()
             except exceptions.CommandTimeout:
                 pass
             except Exception as ex:  # pragma: no cover
@@ -153,7 +153,7 @@ async def check_status(request: web.Request) -> web.Response:
 @routes.post('/system/ping')
 async def ping(request: web.Request) -> web.Response:
     return web.json_response(
-        await device.get_device(request.app).noop()
+        await spark.fget(request.app).noop()
     )
