+++ brewblox_devcon_spark/codec/__init__.py
@@ -5,13 +5,13 @@ Default exports for codec module
 from aiohttp import web
 
 from . import codec, unit_conversion
-from .codec import Codec, TranscodeFunc_, get_codec
+from .codec import Codec, TranscodeFunc_, fget
 from .opts import CodecOpts, FilterOpt, MetadataOpt, ProtoEnumOpt
 from .unit_conversion import get_converter
 
 __all__ = [
     'Codec',
-    'get_codec',
+    'fget',
     'get_converter',
     'setup',
     'CodecOpts',
