+++ test/test_integration.py
@@ -13,8 +13,8 @@ import pytest
 from brewblox_service import mqtt, scheduler
 from brewblox_service.testing import response
 
-from brewblox_devcon_spark import (commander, communication, datastore, device,
-                                   service_status, simulator, synchronization)
+from brewblox_devcon_spark import (commander, connection, datastore, device,
+                                   service_status, simulator, synchronizer)
 from brewblox_devcon_spark.__main__ import parse_ini
 from brewblox_devcon_spark.api import (blocks_api, debug_api, error_response,
                                        settings_api, system_api)
@@ -25,7 +25,7 @@ DEVICE_ID = '123456789012345678901234'
 
 @pytest.fixture(scope='module', autouse=True)
 def firmware_sim():
-    with patch(communication.__name__ + '.BASE_RETRY_INTERVAL_S', 0.1):
+    with patch(connection.__name__ + '.BASE_RETRY_INTERVAL_S', 0.1):
         sim = simulator.FirmwareSimulator()
         sim.start(DEVICE_ID)
         yield
@@ -46,7 +46,7 @@ def app(app):
 
     service_status.setup(app)
 
-    communication.setup(app)
+    connection.setup(app)
     commander.setup(app)
 
     scheduler.setup(app)
@@ -63,7 +63,7 @@ def app(app):
     system_api.setup(app)
     settings_api.setup(app)
 
-    synchronization.setup(app)
+    synchronizer.setup(app)
     return app
