+++ test/test_synchronizer.py
@@ -1,5 +1,5 @@
 """
-Tests brewblox_devcon_spark.synchronization
+Tests brewblox_devcon_spark.synchronizer
 """
 
 import asyncio
@@ -9,11 +9,11 @@ from brewblox_service import brewblox_logger, repeater, scheduler
 from mock import AsyncMock
 
 from brewblox_devcon_spark import (commander_sim, datastore, device,
-                                   service_status, synchronization)
+                                   service_status, synchronizer)
 from brewblox_devcon_spark.codec import codec, unit_conversion
 from brewblox_devcon_spark.service_status import StatusDescription
 
-TESTED = synchronization.__name__
+TESTED = synchronizer.__name__
 LOGGER = brewblox_logger(__name__)
 
 
@@ -27,8 +27,8 @@ def states(app):
 
 
 async def connect(app):
-    service_status.set_connected(app, 'synchronization test')
-    await synchronization.get_syncher(app).sync_done.wait()
+    service_status.set_connected(app, 'synchronizer test')
+    await synchronizer.fget(app).sync_done.wait()
     await asyncio.sleep(0.01)
 
 
@@ -69,18 +69,8 @@ async def app(app, loop):
     unit_conversion.setup(app)
     codec.setup(app)
     device.setup(app)
-    synchronization.setup(app)
+    synchronizer.setup(app)
     return app
-
-
-@pytest.fixture
-async def block_store(app, loop):
-    return datastore.get_block_store(app)
-
-
-@pytest.fixture
-def service_store(app):
-    return datastore.get_config_store(app)
 
 
 @pytest.fixture(autouse=True)
@@ -93,7 +83,7 @@ def api_mock(mocker):
 
 @pytest.fixture
 def syncher(app):
-    return synchronization.get_syncher(app)
+    return synchronizer.fget(app)
 
 
 async def test_sync_status(app, client, mocker):
@@ -121,7 +111,7 @@ async def test_sync_errors(app, client, mocker, system_exit_mock):
 
     assert states(app) == [False, True, False]
     assert system_exit_mock.call_count == 1
-    assert not synchronization.get_syncher(app).active
+    assert not synchronizer.fget(app).active
 
 
 async def test_write_error(app, client, mocker, api_mock, system_exit_mock):
@@ -132,7 +122,7 @@ async def test_write_error(app, client, mocker, api_mock, system_exit_mock):
 
     assert states(app) == [False, True, False]
     assert system_exit_mock.call_count == 1
-    assert not synchronization.get_syncher(app).active
+    assert not synchronizer.fget(app).active
 
 
 async def test_timeout(app, client, syncher, mocker, system_exit_mock):
